+++ generators/spring-service/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable consistent-return */
 const chalk = require('chalk');
 const SpringServiceGenerator = require('generator-jhipster/generators/spring-service');
-const writeFiles = require('./files').writeFiles;
+// const writeFiles = require('./files').writeFiles;
 
 module.exports = class extends SpringServiceGenerator {
     constructor(args, opts) {
@@ -69,6 +69,7 @@ module.exports = class extends SpringServiceGenerator {
 
     get writing() {
         // The writing phase is being overriden so that we can write our own templates as well.
-        return writeFiles();
+        // return writeFiles();
+        return null;
     }
 };